Adding up and down database migration scripts
