--- backend/src/migrate-to-latest.ts
@@ -1,46 +1,0 @@
-require('dotenv').config()
-import * as path from 'path'
-import { Pool } from 'pg'
-import { promises as fs } from 'fs'
-import {
-  Kysely,
-  Migrator,
-  PostgresDialect,
-  FileMigrationProvider,
-} from 'kysely'
-
-import servicelocator from './ServiceLocator'
-
-async function migrateToLatest() {
-  const db = servicelocator.database()
-
-  const migrator = new Migrator({
-    db,
-    provider: new FileMigrationProvider({
-      fs,
-      path,
-      // This needs to be an absolute path.
-      migrationFolder: path.join(__dirname, './Migrations'),
-    }),
-  })
-
-  const { error, results } = await migrator.migrateToLatest()
-
-  results?.forEach((it) => {
-    if (it.status === 'Success') {
-      console.log(`migration "${it.migrationName}" was executed successfully`)
-    } else if (it.status === 'Error') {
-      console.error(`failed to execute migration "${it.migrationName}"`)
-    }
-  })
-
-  if (error) {
-    console.error('failed to migrate')
-    console.error(error)
-    process.exit(1)
-  }
-
-  await db.destroy()
-}
-
-migrateToLatest()
